backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database and caching
redis==5.0.1

# LLM inference
llama-cpp-python==0.2.20

# CORS and middleware
python-dotenv==1.0.0

# System monitoring
psutil==5.9.6

# Additional utilities
aiofiles==23.2.1
websockets==12.0
